Converted queries to accept Integer values

+++ AccountSystemTranslator/src/main/java/za/ac/nwu/ac/translator/MemberAccountTranslator.java
@@ -7,5 +7,5 @@ public interface MemberAccountTranslator {
 
     MemberAccountDto getMember(Long memberID, Long accountTypeID);
 
-    MemberAccountDto updateMemberAccount(Long amount, Long memberId, Long accountTypeId);
+    MemberAccountDto updateMemberAccount(Integer amount, Long memberId, Long accountTypeId);
 }
